LLM-Projects/Conversational_Pdf_ChatBot/requirements.txt
langchain
streamlit
langchain-huggingface
python-dotenv
langchain_community
faiss-gpu